# project-root/docs/04_api/swagger-draft.yaml
openapi: 3.0.3
info:
  title: 个性化旅游系统 API
  version: 1.1.0-draft
  description: |
    个性化旅游系统后端接口文档（Swagger / OpenAPI 3.0）。
    覆盖用户认证、偏好设置、目的地推荐、路线规划、周边设施、美食推荐、
    旅游日记、文件上传、管理端基础维护、AI 增强能力等接口。
    AI 接口属于 P2 增强能力，当前阶段可暂不实现，可先使用 mock / stub。
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
...
  /destinations/recommend:
    get:
      tags: [Destination]
      summary: 获取推荐目的地列表
      security: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [scenic, campus]
        - name: theme
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [heat, rating, recommend]
        - name: pageNum
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: topK
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageDestinationVO'
  /destinations/search:
    get:
      tags: [Destination]
      summary: 搜索目的地
      security: []
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [scenic, campus]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [heat, rating]
        - name: pageNum
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 搜索成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageDestinationVO'
  /destinations/{id}:
    get:
      tags: [Destination]
      summary: 获取目的地详情
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 获取详情成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDestinationVO'
  /destinations/{id}/places:
    get:
      tags: [Destination]
      summary: 获取目的地下的场所列表
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: placeType
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 获取场所列表成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePlaceVOList'
  /destinations/{id}/diaries:
    get:
      tags: [Destination]
      summary: 获取目的地相关日记列表
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [heat, rating, latest]
        - name: pageNum
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageDiaryVO'
...
  /foods/recommend:
    get:
      tags: [Food]
      summary: 获取美食推荐列表
      security: []
      parameters:
        - name: destinationId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: sourceNodeId
          in: query
          schema:
            type: integer
            format: int64
        - name: foodType
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [heat, rating, distance]
        - name: topK
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageFoodVO'
  /foods/search:
    get:
      tags: [Food]
      summary: 搜索美食
      security: []
      parameters:
        - name: destinationId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [heat, rating, distance]
        - name: pageNum
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 搜索成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageFoodVO'
  /destinations/{id}/comments:
    get:
      tags: [Destination]
      summary: 分页查看目的地一级评论
      security: []
      parameters:
        - $ref: '#/components/parameters/CommentTargetId'
        - $ref: '#/components/parameters/CommentPageNum'
        - $ref: '#/components/parameters/CommentPageSize'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageCommentVO'
    post:
      tags: [Destination]
      summary: 发布目的地一级评论
      parameters:
        - $ref: '#/components/parameters/CommentTargetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCommentVO'
  /foods/{id}/comments:
    get:
      tags: [Food]
      summary: 分页查看美食一级评论
      security: []
      parameters:
        - $ref: '#/components/parameters/CommentTargetId'
        - $ref: '#/components/parameters/CommentPageNum'
        - $ref: '#/components/parameters/CommentPageSize'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageCommentVO'
    post:
      tags: [Food]
      summary: 发布美食一级评论
      parameters:
        - $ref: '#/components/parameters/CommentTargetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCommentVO'
  /diaries:
    post:
      tags: [Diary]
      summary: 发布日记
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDiaryRequest'
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCreateDiaryResult'
    get:
      tags: [Diary]
      summary: 获取日记列表
      security: []
      parameters:
        - name: destinationId
          in: query
          schema:
            type: integer
            format: int64
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [heat, rating, latest]
        - name: pageNum
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取列表成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageDiaryVO'
...
  /diaries/{id}:
    get:
      tags: [Diary]
      summary: 获取日记详情
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 获取详情成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDiaryVO'
  /diaries/search/title:
    get:
      tags: [Diary]
      summary: 按标题精确查询日记
      security: []
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageDiaryVO'
  /diaries/search/fulltext:
    get:
      tags: [Diary]
      summary: 日记全文检索
      security: []
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - name: destinationId
          in: query
          schema:
            type: integer
            format: int64
        - name: pageNum
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 检索成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageDiaryVO'
  /diaries/{id}/ratings:
    post:
      tags: [Diary]
      summary: 对日记评分
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiaryRatingRequest'
      responses:
        '200':
          description: 评分成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseBoolean'
  /diaries/{id}/ratings/me:
    get:
      tags: [Diary]
      summary: 获取当前用户对日记的评分
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDiaryRatingVO'
  /diaries/{id}/comments:
    get:
      tags: [Diary]
      summary: 分页查看日记一级评论
      security: []
      parameters:
        - $ref: '#/components/parameters/CommentTargetId'
        - $ref: '#/components/parameters/CommentPageNum'
        - $ref: '#/components/parameters/CommentPageSize'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageCommentVO'
    post:
      tags: [Diary]
      summary: 发布日记一级评论
      parameters:
        - $ref: '#/components/parameters/CommentTargetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCommentVO'
...
  /admin/destinations/{id}:
    put:
      tags: [Admin]
      summary: 修改目的地
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminDestinationUpsertRequest'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseBoolean'
    delete:
      tags: [Admin]
      summary: 删除目的地
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseBoolean'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CommentTargetId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    CommentPageNum:
      name: pageNum
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    CommentPageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
    ApiResponseBase:
      type: object
      properties:
        success:
          type: boolean
          example: true
        code:
          type: string
          example: SUCCESS
        message:
          type: string
          example: success
        timestamp:
          type: string
          format: date-time
    ApiResponseBoolean:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: boolean
              example: true
...
    ApiResponseDestinationVO:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/DestinationVO'
    ApiResponsePlaceVOList:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/PlaceVO'
...
    ApiResponsePageFoodVO:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageFoodVO'
    ApiResponseDiaryVO:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/DiaryVO'
    ApiResponseDiaryRatingVO:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/DiaryRatingVO'
    ApiResponseCommentVO:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CommentVO'
    ApiResponsePageCommentVO:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageCommentVO'
    ApiResponseCreateDiaryResult:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              type: object
              properties:
                diaryId:
                  type: integer
                  format: int64
...
    ApiResponsePageDestinationVO:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageDestinationVO'
    ApiResponsePageDiaryVO:
      allOf:
        - $ref: '#/components/schemas/ApiResponseBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageDiaryVO'
...
    CreateDiaryRequest:
      type: object
      required: [destinationId, title, contentText]
      properties:
        destinationId:
          type: integer
          format: int64
        routeHistoryId:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
          maxLength: 150
        contentText:
          type: string
        visibility:
          type: string
          enum: [public, private]
          default: public
        mediaIds:
          type: array
          items:
            type: integer
            format: int64
    DiaryRatingRequest:
      type: object
      required: [score]
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 5
    DiaryRatingVO:
      type: object
      properties:
        diaryId:
          type: integer
          format: int64
        userScore:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        ratingScore:
          type: number
          format: double
        ratingCount:
          type: integer
          minimum: 0
    CommentCreateRequest:
      type: object
      required: [contentText]
      properties:
        contentText:
          type: string
          minLength: 1
          maxLength: 500
    CommentVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        targetType:
          type: string
          enum: [destination, food, diary]
        targetId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        nickname:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        contentText:
          type: string
        createdAt:
          type: string
          format: date-time
    PageCommentVO:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/CommentVO'
        pageNum:
          type: integer
          format: int64
        pageSize:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
        pages:
          type: integer
          format: int64
    AdminDestinationUpsertRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [scenic, campus]
        category:
          type: string
        city:
          type: string
        description:
          type: string
        heatScore:
          type: number
          format: double
        ratingScore:
          type: number
          format: double
        coverUrl:
          type: string
        tags:
          type: array
          items:
            type: string
...
    DestinationVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
        category:
          type: string
        city:
          type: string
        description:
          type: string
        heatScore:
          type: number
          format: double
        ratingScore:
          type: number
          format: double
        coverUrl:
          type: string
        tags:
          type: array
          items:
            type: string
    PlaceVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        destinationId:
          type: integer
          format: int64
        name:
          type: string
        placeType:
          type: string
        description:
          type: string
        lng:
          type: number
          format: double
        lat:
          type: number
          format: double
        floorInfo:
          type: string
        openTimeRule:
          type: string
        suggestedDurationMin:
          type: integer
        costLevel:
          type: integer
...
    FoodVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        destinationId:
          type: integer
          format: int64
        facilityId:
          type: integer
          format: int64
        name:
          type: string
        foodType:
          type: string
        shopName:
          type: string
        description:
          type: string
        heatScore:
          type: number
          format: double
        ratingScore:
          type: number
          format: double
        avgPrice:
          type: number
          format: double
        coverUrl:
          type: string
        lng:
          type: number
          format: double
        lat:
          type: number
          format: double
    DiaryMediaVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        mediaType:
          type: string
          enum: [image, video]
        fileUrl:
          type: string
    DiaryVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        username:
          type: string
        destinationId:
          type: integer
          format: int64
        destinationName:
          type: string
        routeHistoryId:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
        contentText:
          type: string
        heatScore:
          type: number
          format: double
        ratingScore:
          type: number
          format: double
        ratingCount:
          type: integer
          minimum: 0
        visibility:
          type: string
          enum: [public, private]
        mediaList:
          type: array
          items:
            $ref: '#/components/schemas/DiaryMediaVO'
        createdAt:
          type: string
          format: date-time
...
    PageDestinationVO:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/DestinationVO'
        pageNum:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
        pages:
          type: integer
...
    PageFoodVO:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/FoodVO'
        pageNum:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
        pages:
          type: integer
    PageDiaryVO:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/DiaryVO'
        pageNum:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
        pages:
          type: integer
...